move detectors into folders
