+++ source-code/sdk-js/src/detectors/detectors/client/navigatorDetector.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test, vi } from "vitest"
-import { navigatorDetectorTemplate } from "./navigatorDetector.js"
+import { navigatorDetector } from "./navigatorDetector.js"
 
 const langs = ["en", "de", "en-GB"]
 
@@ -11,7 +11,7 @@ vi.stubGlobal("window", {
 
 describe("navigatorDetector", () => {
 	test("returns languages from window object", () => {
-		const detector = navigatorDetectorTemplate
+		const detector = navigatorDetector
 		expect(detector()).toEqual(langs)
 	})
 })
