+++ source-code/sdk-js/src/detectors/detectors/server/acceptLanguageHeaderDetector.test.ts
