+++ source-code/sdk-js/src/detectors/index.ts
@@ -1,8 +1,9 @@
+export type { Detector, DetectorInitializer } from "./types.js"
 export { detectLanguage } from "./detectLanguage.js"
+
 export {
 	initAcceptLanguageHeaderDetector,
 	acceptLanguageHeaderDetector,
-} from "./acceptLanguageHeaderDetector.js"
-export { navigatorDetectorTemplate } from "./navigatorDetector.js"
-export { initRootSlugDetector, rootSlugDetectorTemplate } from "./rootSlugDetector.js"
-export type { Detector, DetectorInitializer } from "./types.js"
+} from "./detectors/server/acceptLanguageHeaderDetector.js"
+export { navigatorDetector } from "./detectors/client/navigatorDetector.js"
+export { initRootSlugDetector, rootSlugDetector } from "./detectors/shared/rootSlugDetector.js"
